+++ src/Data/index.ts
@@ -1,4 +1,4 @@
-export * from "./CardData";
+export * from "./Cards";
 export * from "./Constants";
 export * from "./Decklists";
 export * from "./Enums";