+++ docs/RELEASE-NOTES.md
@@ -1,3 +1,3 @@
-[1.0.1] EnterDecklistPage Dropdown
-- Minor css improvements
-- Beijing Open
+[1.1.0] Backend tweaks
+- JSON and python scripts for card info and id mappings
+- More events